requirements.txt
# ArgusHR Backend Requirements

# Core
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0

# RAG Pipeline
voyageai>=0.2.0
chromadb>=0.4.0
tiktoken>=0.5.0
groq>=0.4.0
supabase>=2.0.0

# Document Processing
pandas>=2.0.0
pdfplumber>=0.10.0
openpyxl>=3.1.0
tabulate>=0.9.0

# Utilities
numpy>=1.24.0
pydantic>=2.0.0
